placement-react-drf-mysql/placement-react-drf-mysql/backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
dj-database-url>=2.2
